modified testing strategy
